requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
pytest==7.4.3
python-dotenv==1.0.0
WTForms==3.1.1
email-validator==2.1.0.post1 
psycopg2-binary==2.9.9